+++ CHANGELOG.md
@@ -13,8 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Update indicator arrow flickers when opening the popup [#3351].
 - $badfilter incorrectly negates rules with different $denyallow values [#3428].
 
+[#3351]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3351
 [#3330]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3330
 [#3428]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3428